requirements.txt
openai
pyyaml
requests
numpy
